Revert "idk maybe it will work"
This reverts commit f114ad793822e1dafc07dd69d1ba1518cc901008.

+++ ScriptedEvents/Actions/Player/EffectAction.cs
@@ -77,7 +77,7 @@
 
                     foreach (Player player in plys)
                     {
-                        player.ChangeEffectIntensity(effect, intensity);
+                        player.ChangeEffectIntensity(effect, (byte)intensity);
                         player.EnableEffect(effect, duration);
                     }
